+++ main_penalty_calculate/builders/__init__.py
@@ -1,2 +1,3 @@
 from .violators_avaliations_builder import ViolatorsAvaliationsBuilder
 from .traffic_violation_builder import TrafficViolationBuilder
+from .penalties_values_builder import PenaltiesValuesBuilder
