+++ main_penalty_calculate/models/__init__.py
@@ -2,4 +2,3 @@ from .identity_card import IdentityCard
 from .license_plate import LicensePlate
 from .traffic_violation import TrafficViolation
 from .violator_avaliation import ViolatorAvaliation
-from .penaltis_values_builder import PenaltiesValuesBuilder
